Rewrote wish sending with okhttp
